% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amd_plots.R
\name{amd_plots}
\alias{amd_plots}
\title{Daily, monthly and annual time series plots}
\usage{
amd_plots(data, date, rainfall)
}
\arguments{
\item{data}{A file path of the csv file as a string within double quotes. For example use Prec_data.csv demo file using system.file()}

\item{date}{Column name which contain date and year as a string within double quotes. Date has to be in year-month-day format.}

\item{rainfall}{column name which contain rainfall/precipitation values as string within double quotes.}
}
\value{
9 plots
}
\description{
The function "amd_plots" can be used to plot time series plots for rainfall data.
Daily/monthly/annual/seasonal rainfall values of the time series given as input. Depending on the value of daily, monthly, annual and/or seasonal time series plots, boxplots and histograms are produced.
}
\examples{
file= system.file("extdata","Prec_data.csv", package = "Rwplots")
amd_plots(file,"time","Rainfall")
}
