% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_plot.R
\name{matrix_plot}
\alias{matrix_plot}
\title{Matrix plot}
\usage{
matrix_plot(data, date, rainfall)
}
\arguments{
\item{data}{A file path of the csv file as a string within double quotes. For example use Prec_data.csv demo file using system.file()}

\item{date}{Column name which contain date and year as a string within double quotes. Date has to be in year-month-day format.}

\item{rainfall}{column name which contain rainfall/precipitation values as string within double quotes.}
}
\value{
a plot
}
\description{
This function "matrix_plot" is used to convert values to matrix and plot the precipitation values.
}
\examples{
file= system.file("extdata","Prec_data.csv", package = "Rwplots")
matrix_plot(file,"time","Rainfall")
}
