% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_plot.R
\name{precip_plot}
\alias{precip_plot}
\title{Annual precipitation plot}
\usage{
precip_plot(data, rainfall, lon, lat)
}
\arguments{
\item{data}{A file path of the csv file as a string within double quotes. For example use Rainfall.nc demo file using system.file()}

\item{rainfall}{Column name which contain Rainfall values as a string within double quotes}

\item{lon}{Column name which contain longitude values as a string within double quotes}

\item{lat}{Column name which contain latitude values as a string within double quotes}
}
\value{
A precipitation plot
}
\description{
The function "Precip_plot" can be used to plot annual precipitation plot with nc file. The resulting plot will looks like large when you see it in the plots tab so click zoom to see it in extract size.
}
\examples{
file= system.file("extdata","Rainfall.nc", package = "Rwplots")
precip_plot(file,"RAINFALL","LONGITUDE", "LATITUDE")
}
