% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackbyyear.R
\name{trackbyyear}
\alias{trackbyyear}
\title{Yearwise cyclones tracks plotting}
\usage{
trackbyyear(data, lat, lon, wind, name, year, val, reg)
}
\arguments{
\item{data}{A file path of the csv file as a string within double quotes. For example use Tracks.csv demo file using system.file()}

\item{lat}{Column name which contain latitude values as a string within double quotes}

\item{lon}{Column name which contain longitude values as a string within double quotes}

\item{wind}{Column name which contain wind values as a string within double quotes}

\item{name}{Column name which contain cyclone name values as a string within double quotes}

\item{year}{Column name which contain only year as a string within double quotes}

\item{val}{specific year which you want plot as numeric value}

\item{reg}{The region name around which your data has to be plotted. For examples "India".}
}
\value{
A plot of particular year
}
\description{
The function "trackbyyear" can be used to plot cyclone tracks for particular year and also uses different kind of plotting methods rather the one which is used in ts_tracks function. You can also use an external data in this function which should have all the required columns(refer arguments).
If you face some error in filter then try rename the columns name exactly like given in the example.
}
\examples{
file= system.file("extdata","Tracks.csv", package = "Rwplots")
trackbyyear(file,"lat","lon", "kts", "name","year",2018,"India")
}
