% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_tracks.R
\name{ts_tracks}
\alias{ts_tracks}
\title{Time series plot of cyclone tracks}
\usage{
ts_tracks(data, Lon, Lat, Name, Year)
}
\arguments{
\item{data}{A file path of the csv file as a string. For example use Tracks.csv demo file using system.file().}

\item{Lon}{Column name which contain longitude values as a string within double quotes}

\item{Lat}{Column name which contain latitude values as a string within double quotes}

\item{Name}{Column name which contain cyclone name values as a string within double quotes}

\item{Year}{Column name which contain only year as a string within double quotes}
}
\value{
a plot
}
\description{
The function "ts_tracks" can be used to plot  time series of cyclone tracks for the total number of years according to the given data.
}
\examples{
file= system.file("extdata","Tracks.csv", package = "Rwplots")
ts_tracks(file,"lon","lat","name","year")
}
